future(homepage): export widget components
