+++ packages/core/admin/admin/src/index.ts
@@ -17,6 +17,7 @@ export * from './components/Filters';
 export * from './components/Form';
 export * from './components/FormInputs/Renderer';
 export * from './components/PageHelpers';
+export * from './components/WidgetHelpers';
 export * from './components/Pagination';
 export * from './components/SearchInput';
 export * from './components/Table';
