+++ packages/core/admin/admin/src/pages/Home/HomePage.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 
-import { Flex, Grid, Main } from '@strapi/design-system';
-import { useIntl } from 'react-intl';
+import { Box, Flex, Grid, Main, Typography } from '@strapi/design-system';
+import { CheckCircle, Pencil, PuzzlePiece } from '@strapi/icons';
+import { MessageDescriptor, useIntl } from 'react-intl';
 
 import { Layouts } from '../../components/Layouts/Layout';
 import { Page } from '../../components/PageHelpers';
@@ -10,6 +11,47 @@ import { useAuth } from '../../features/Auth';
 
 import { LastEditedWidget, LastPublishedWidget } from './components/ContentManagerWidgets';
 import { GuidedTour } from './components/GuidedTour';
+
+/* -------------------------------------------------------------------------------------------------
+ * Root
+ * -----------------------------------------------------------------------------------------------*/
+
+interface RootProps {
+  title: MessageDescriptor;
+  icon?: typeof import('@strapi/icons').PuzzlePiece;
+  children: React.ReactNode;
+}
+const WidgetRoot = ({ title, icon = PuzzlePiece, children }: RootProps) => {
+  const { formatMessage } = useIntl();
+  const id = React.useId();
+  const Icon = icon;
+
+  return (
+    <Flex
+      width="100%"
+      hasRadius
+      direction="column"
+      alignItems="flex-start"
+      background="neutral0"
+      borderColor="neutral150"
+      shadow="tableShadow"
+      tag="section"
+      gap={4}
+      padding={6}
+      aria-labelledby={id}
+    >
+      <Flex direction="row" alignItems="center" gap={2} tag="header">
+        <Icon fill="neutral500" aria-hidden />
+        <Typography textColor="neutral500" variant="sigma" tag="h2" id={id}>
+          {formatMessage(title)}
+        </Typography>
+      </Flex>
+      <Box width="100%" height="261px" overflow="auto" tag="main">
+        {children}
+      </Box>
+    </Flex>
+  );
+};
 
 /* -------------------------------------------------------------------------------------------------
  * HomePageCE
@@ -40,10 +82,26 @@ const HomePageCE = () => {
           <GuidedTour />
           <Grid.Root gap={5}>
             <Grid.Item col={6} s={12}>
-              <LastEditedWidget />
+              <WidgetRoot
+                title={{
+                  id: 'content-manager.widget.last-edited.title',
+                  defaultMessage: 'Last edited entries',
+                }}
+                icon={Pencil}
+              >
+                <LastEditedWidget />
+              </WidgetRoot>
             </Grid.Item>
             <Grid.Item col={6} s={12}>
-              <LastPublishedWidget />
+              <WidgetRoot
+                title={{
+                  id: 'content-manager.widget.last-published.title',
+                  defaultMessage: 'Last published entries',
+                }}
+                icon={CheckCircle}
+              >
+                <LastPublishedWidget />
+              </WidgetRoot>
             </Grid.Item>
           </Grid.Root>
         </Flex>
